requirements.txt
langchain>=0.2.10
langchain-openai>=0.1.7
openai>=1.40.0
streamlit>=1.36.0
python-dotenv>=1.0.1
pydantic>=2.7.4

